+++ src/main/java/insane96mcp/enhancedai/modules/zombie/feature/DiggerZombie.java
@@ -8,9 +8,9 @@ import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.config.BlacklistConfig;
 import insane96mcp.insanelib.util.IdTagMatcher;
-import net.minecraft.world.entity.ai.goal.target.NearestAttackableTargetGoal;
-import net.minecraft.world.entity.monster.Endermite;
 import net.minecraft.world.entity.monster.Zombie;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
 import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.event.entity.EntityJoinWorldEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -23,6 +23,7 @@ public class DiggerZombie extends Feature {
 	private final ForgeConfigSpec.ConfigValue<Double> diggerChanceConfig;
 	private final ForgeConfigSpec.ConfigValue<Boolean> diggerToolOnlyConfig;
 	private final ForgeConfigSpec.ConfigValue<Boolean> diggerProperToolOnlyConfig;
+	private final ForgeConfigSpec.ConfigValue<Boolean> equipWoodenPickConfig;
 	private final ForgeConfigSpec.ConfigValue<Boolean> blacklistTileEntitiesConfig;
 	private final ForgeConfigSpec.ConfigValue<Double> miningSpeedMultiplierConfig;
 	private final BlacklistConfig blockBlacklistConfig;
@@ -31,8 +32,9 @@ public class DiggerZombie extends Feature {
 	public double diggerChance = 0.05;
 	public boolean diggerToolOnly = false;
 	public boolean diggerProperToolOnly = false;
-	public double miningSpeedMultiplier = 1d;
+	public boolean equipWoodenPick = true;
 	public boolean blacklistTileEntities = false;
+	public double miningSpeedMultiplier = 1d;
 	public ArrayList<IdTagMatcher> blockBlacklist;
 	public boolean blockBlacklistAsWhitelist;
 	public ArrayList<IdTagMatcher> entityBlacklist;
@@ -51,6 +53,9 @@ public class DiggerZombie extends Feature {
 		diggerProperToolOnlyConfig = Config.builder
 				.comment("Zombies with Digger AI will mine only if their off-hand tool can mine targeted blocks (e.g. zombies with shovels will not mine stone). Blocks that require no tool (e.g. planks) will be minable.")
 				.define("Digger Proper Tool Only", this.diggerProperToolOnly);
+		equipWoodenPickConfig = Config.builder
+				.comment("Zombies with Digger AI will spawn with a Wooden Pickaxe.")
+				.define("Equip Wooden Pick", this.equipWoodenPick);
 		miningSpeedMultiplierConfig = Config.builder
 				.comment("Multiplier for digger zombies mining speed. E.g. with this set to 2, zombies will take twice the time to mine a block.")
 				.defineInRange("Digger Speed Multiplier", this.miningSpeedMultiplier, 0d, 128d);
@@ -68,6 +73,7 @@ public class DiggerZombie extends Feature {
 		this.diggerChance = this.diggerChanceConfig.get();
 		this.diggerToolOnly = this.diggerToolOnlyConfig.get();
 		this.diggerProperToolOnly = this.diggerProperToolOnlyConfig.get();
+		this.equipWoodenPick = this.equipWoodenPickConfig.get();
 		this.miningSpeedMultiplier = this.miningSpeedMultiplierConfig.get();
 		this.blacklistTileEntities = this.blacklistTileEntitiesConfig.get();
 		this.blockBlacklist = (ArrayList<IdTagMatcher>) IdTagMatcher.parseStringList(this.blockBlacklistConfig.listConfig.get());
@@ -111,9 +117,9 @@ public class DiggerZombie extends Feature {
 			zombie.getPersistentData().putBoolean(Strings.Tags.PROCESSED, true);
 		}
 
-		if (miner)
+		if (miner) {
 			zombie.goalSelector.addGoal(1, new DiggingGoal(zombie, this.diggerToolOnly, this.diggerProperToolOnly));
-
-		zombie.targetSelector.addGoal(2, new NearestAttackableTargetGoal<>(zombie, Endermite.class, true));
+			zombie.equipItemIfPossible(new ItemStack(Items.WOODEN_PICKAXE));
+		}
 	}
 }